requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
bandit==1.9.4
blinker==1.9.0
boolean-py==5.0
cachecontrol==0.14.4
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
discord-py==2.7.1
filelock==3.25.1
flask==3.1.3
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
license-expression==30.4.4
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
packageurl-python==0.17.6
packaging==26.0
pip==26.0.1
pip-api==0.0.34
pip-audit==2.10.0
pip-requirements-parser==32.0.1
platformdirs==4.9.4
propcache==0.4.1
py-serializable==2.1.0
pygments==2.19.2
pyparsing==3.3.2
pyyaml==6.0.3
requests==2.32.5
rich==14.3.3
ruff==0.15.5
sortedcontainers==2.4.0
stevedore==5.7.0
tomli==2.4.0
tomli-w==1.2.0
urllib3==2.6.3
werkzeug==3.1.6
yarl==1.23.0
